fix: Refactor testing and sort out unit and integration tests (#2975)
* Refactor go feature server


* Fix lint


* Fix e2e tests


* Fix


* Verify tests


* Fix lint


* Address review


* Fix


* Address review


* Fix


* Fix lint


* address review


* Fix


* Fix


* Fix lint


* Fix lint


* Fix


* Fix


* Fix lint


* Fix


* Refactor


* Fix


* Fx lit


* Fix lint


* update fix


* Fix


* Revert


* Fix


* fix


* Fix lint
