+++ sdk/python/tests/integration/e2e/test_usage_e2e.py
@@ -11,6 +11,8 @@
 # WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
 # See the License for the specific language governing permissions and
 # limitations under the License.
+
+# This file tests our usage tracking system in `usage.py`.
 import os
 import sys
 import tempfile
