+++ sdk/python/tests/unit/local_feast_tests/test_feature_service_read.py
@@ -1,7 +1,7 @@
 import pytest
 
-from tests.utils.cli_utils import CliRunner, get_example_repo
-from tests.utils.online_read_write_test import basic_rw_test
+from tests.utils.basic_read_write_test import basic_rw_test
+from tests.utils.cli_repo_creator import CliRunner, get_example_repo
 
 
 @pytest.mark.integration
