+++ sdk/python/tests/unit/local_feast_tests/test_init.py
@@ -3,7 +3,7 @@ from datetime import datetime, timedelta
 from pathlib import Path
 from textwrap import dedent
 
-from tests.utils.cli_utils import CliRunner
+from tests.utils.cli_repo_creator import CliRunner
 
 
 def test_repo_init() -> None:
