+++ sdk/python/tests/utils/test_log_creator.py
@@ -3,16 +3,74 @@ import datetime
 import tempfile
 import uuid
 from pathlib import Path
-from typing import Iterator, Union
+from typing import Iterator, List, Union
 
 import numpy as np
 import pandas as pd
 import pyarrow
+import pytz
 
 from feast import FeatureService, FeatureStore, FeatureView
 from feast.errors import FeatureViewNotFoundException
 from feast.feature_logging import LOG_DATE_FIELD, LOG_TIMESTAMP_FIELD, REQUEST_ID_FIELD
 from feast.protos.feast.serving.ServingService_pb2 import FieldStatus
+
+
+def get_latest_rows(
+    df: pd.DataFrame, join_key: str, entity_values: List[str]
+) -> pd.DataFrame:
+    """
+    Return latest rows in a dataframe based on join key and entity values.
+
+    Args:
+        df: Dataframe of features values.
+        join_key : Join key for the feature values in the dataframe.
+        entity_values : Entity values for the feature values in the dataframe.
+
+    Returns:
+        The most recent row in the dataframe.
+    """
+    rows = df[df[join_key].isin(entity_values)]
+    return rows.loc[rows.groupby(join_key)["event_timestamp"].idxmax()]
+
+
+def generate_expected_logs(
+    df: pd.DataFrame,
+    feature_view: FeatureView,
+    features: List[str],
+    join_keys: List[str],
+    timestamp_column: str,
+) -> pd.DataFrame:
+    """
+    Given dataframe and feature view, generate the expected logging dataframes that would be otherwise generated by our logging infrastructure.
+    Args:
+        df: Dataframe of features values returned in `get_online_features`.
+        feature_view : The feature view from which the features were retrieved.
+        features : The list of features defined as part of this base feature view.
+        join_keys : Join keys for the retrieved features.
+        timestamp_column : Timestamp column
+
+    Returns:
+        Returns dataframe containing the expected logs.
+    """
+    logs = pd.DataFrame()
+    for join_key in join_keys:
+        logs[join_key] = df[join_key]
+
+    for feature in features:
+        col = f"{feature_view.name}__{feature}"
+        logs[col] = df[feature]
+        logs[f"{col}__timestamp"] = df[timestamp_column]
+        logs[f"{col}__status"] = FieldStatus.PRESENT
+        if feature_view.ttl:
+            logs[f"{col}__status"] = logs[f"{col}__status"].mask(
+                df[timestamp_column]
+                < datetime.datetime.utcnow().replace(tzinfo=pytz.UTC)
+                - feature_view.ttl,
+                FieldStatus.OUTSIDE_MAX_AGE,
+            )
+
+    return logs.sort_values(by=join_keys).reset_index(drop=True)
 
 
 def prepare_logs(
